+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Improve `Object.checksum()` handling of class instances
 * Add default `Symbol.toStringTag` implementation for classes
 * Add `Decimal` classes
+* Make `Object.isPrimitive(arg)` return true for all primitive types
 
 ## 0.8.13 (2023-10-15)